shove counters into a footer component inside of editor
