+++ src/lib/components/Divider.svelte
@@ -1,1 +1,1 @@
-<div class="min-h-[1em] w-[0.1rem] self-stretch bg-neutral-100 bg-opacity-100" />
+<div class="min-h-[1em] w-[1px] self-stretch bg-neutral-300" />
