+++ src/lib/components/editor/footer/index.svelte
@@ -1,0 +1,7 @@
+<script>
+	import Counters from './Counters.svelte';
+</script>
+
+<div>
+	<Counters />
+</div>
